+++ R/globals.R
@@ -9,4 +9,5 @@ utils::globalVariables(c("df_fut","dflong","dfwide","expiry_table","holidaysOil"
                          "ticker","name","SupplyNOPEC","SupplyOPEC","Demand","Inv_Change",
                          "tickers_eia","sd_category","tick.r","imports","exports","demand","supply",
                          "ref.opt.inputs","ref.opt.outputs","prices","LightSweet.yield","LightSweet",
-                         "HeavySour","HeavySour.yield","Element","product","max.prod","info","code"))
+                         "HeavySour","HeavySour.yield","Element","product","max.prod","info","code",
+                         "Close","fut.contract"))